added functional for for dividing the personal professional section

+++ my-app/src/components/appList/AppList.js
@@ -19,21 +19,32 @@ class AppList extends Component {
   }
 
   render() {
-
     return (
       <Container>
         <Row className="main__list list p-3 mb-4">
           <ListGroup as="ul" className="list__inner">
-            {this.state.loading ? this.props.data.map(item => {
-              const { _id, ...itemProps } = item;
+            {this.state.loading ? this.props.data.map(item => { // можно чуть переписать, но пока я епал
+              if (this.props.onChoose && item.profile == 'Personal') {
+                const { _id, ...itemProps } = item;
+                return (
+                  <AppListItem
+                    key={_id}
+                    {...itemProps}
+                    deleteItem={() => this.props.deleteItem(_id)}
+                    checkItem={() => this.props.checkItem(_id)} />
+                )
+              }
 
-              return (
-                <AppListItem
-                  key={_id}
-                  {...itemProps}
-                  deleteItem={() => this.props.deleteItem(_id)}
-                  checkItem={() => this.props.checkItem(_id)} />
-              )
+              if (!this.props.onChoose && item.profile == 'Professional') {
+                const { _id, ...itemProps } = item;
+                return (
+                  <AppListItem
+                    key={_id}
+                    {...itemProps}
+                    deleteItem={() => this.props.deleteItem(_id)}
+                    checkItem={() => this.props.checkItem(_id)} />
+                )
+              }
             }): null}
           </ListGroup>
         </Row>
